BackEnd/requirements.txt
fastapi
pydantic
pymongo
uvicorn
pytest
httpx
motor
python-dotenv